move bundler typing to index
